[ADD] membership_initial_fee
Initial fee for memberships
===========================

Charge an initial fee when a partner is invoiced for the first time with a
member product.

Usage
=====

Define a member product, and select 'Fixed amount' or 'Percentage of the price'
for invoicing an extra charge in the first invoice that is created with this
member product.