[ADD] membership_prorrate
===========================
Prorrate the membership fee
===========================

When you have a periodical fee (like a yearly fee), maybe you want to have
membership only for the selected period, but don't charge the customer the
full fee, only the proportional part. With this module you can do it.

Usage
=====

Go to the membership product, and check "Prorrate" in it. When you create
invoices that includes a membership product by hand or pressing the button
"Buy Membership" from members form.