[FIX] membership_initial_fee: Don't apply initial fee on refund/vendor
If you don't have any membership line, and you create a vendor bill/refund or a
customer refund, the initial fee will be added to them, which is not correct.

The check for adding the initial fee now takes into account the invoice type
for excluding those that are the non desired ones.

TT25327