[ADD] membership_custom_date
This module is needed for memberships that have a custom date range for
every individual.

The normal workflow when clicking on 'Buy Membership' does not involve
specifying the start and end dates. Rather, those dates are calculated
from the membership product. This calculation happens _after_ clicking
on 'Invoice Membership' in the wizard UI.

If we want to include the dates in the wizard UI, we disrupt this
workflow, and override the default behaviour of using the dates
configured in the membership product.

We circumvent this by only breaking the workflow for membership products
of type 'custom', which would be expected. We keep the workflow for
other types.

An alternative method is to calculate the dates from the product
selected in the wizard (onchange), but this is UX-wise incompatible with
the a downstream (non-OCA) requirement that the dates be changed to the
dates specified on another model (also onchange). So depending on which
field you changed last (product or other model); that's the dates you
get. This is strange and unexpected behaviour.

Signed-off-by: Carmen Bianca BAKKER <carmen@coopiteasy.be>