[FIX] membership_extension: Use correct view when interacting with membership product field
`membership` provides a form view `membership_products_form` for
`product.template`. This is a custom view that does not inherit
`product_template_form_view`, and is used in 'Members → Configuration →
Membership Products'.

When creating or opening a product from the affected field, the default
form view `product_template_form_view` is displayed as a pop-up. This is
not desirable.

With this change, the correct view is used, and the correct defaults are
set consistent with `membership.action_membership_products`.

Signed-off-by: Carmen Bianca Bakker <carmen@coopiteasy.be>