[IMP] membership_[extension,type,custom_date]: Add restraint on date_from and date_to on membership line
This was a bit tricky to get right, architecturally, but I think I
arrived at the correct design decision. On the whole, the new
restriction makes sense. Following from the fact that dates are
mandatory on 'fixed dates' (read: default) membership products, and
following from the fact that dates are always set on the membership
lines of 'variable period' membership products, membership lines should
_always_ have values for date_from and date_to. The fact that these
fields are not already required is a strange design decision.

The constraint in membership_extension _effectively_ sets required=True
on these fields.

However, there is a use-case for memberships that do not have an end
date (using the 'custom' membership product). For those scenarios, we
add a toggle on the product (exclusive to 'custom' products) on whether
to apply the constraint.

Signed-off-by: Carmen Bianca BAKKER <carmen@coopiteasy.be>